Tweak karma autolearning for Microsoft domains
It seems that, lately, hotmail.com and outlook.com are sending some
spam. This is a problem because the karma autolearning rules make it
so that the spam sent from those domains (which have high karma) ends
up not only not marked as spam, but used as an example of ham to train
the filter.

So, exclude hotmail.com and outlook.com from the heuristic that
ignores spam marks from high-karma domains.